refactor(SE-62): rename `usePropertyDetails` hook to `usePropertyCurrentDetails` for clarity
* Update fetching logic to use `useFetchData`
* Add description indicating it fetches current property and room details, including availability and adjusted price
